+++ packages/theme/docs/.vitepress/blog-theme.ts
@@ -207,7 +207,10 @@ export const blogTheme = getThemeConfig({
     duration: 0
   },
   tabs: true,
-  works: workConfig
+  works: workConfig,
+  footer: {
+    copyright: 'MIT License | 粥里有勺糖',
+  }
 })
 
 export const extraHead: any