+++ packages/theme/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @sugarat/theme
 
+## 0.2.1
+
+### Patch Changes
+
+- chore: update version link
+
+## 0.2.0
+
+### Minor Changes
+
+- feat: 支持首页展示头像
+- feat: 支持更加简单的定义底部footer
+- chore: 默认使用rc17
+
 ## 0.1.50
 
 ### Patch Changes